+++ ChangeLog.md
@@ -3,6 +3,7 @@
 * Add ability to store user state object in DicomClient
 * Fix handling of UIDs in DicomCFindRequest
 * Fix comparison of private DicomTags
+* Add shortcut constructor for private DicomTags
 
 #### v1.0.14 (09/05/2012)
 * Fix bug in DicomDatasetReaderObserver handling zero length Specific Character Set elements